+++ cli/src/commands/mod.rs
@@ -6,7 +6,6 @@ pub mod invite;
 pub mod ls;
 #[cfg(feature = "manpage")]
 pub mod man_page;
-pub mod mount_realm_export;
 pub mod organization;
 pub mod rm;
 #[cfg(feature = "testenv")]
