+++ cli/tests/integration/sequester/mod.rs
@@ -1,1 +1,2 @@
 mod gen_svc_cert;
+mod mount;
