+++ newsfragments/13265.misc.rst
@@ -1,0 +1,1 @@
+[CLI] Renamed command ``mount-realm-export`` to ``sequester mount``
